requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-bootstrap4==1.0.1
django-filter==2.2.0
django-rest-framework==0.1.0
django-stubs==1.1.0
djangorestframework==3.10.2
djangorestframework-stubs==1.0.0
et-xmlfile==1.0.1
httpie==1.0.3
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
MarkupSafe==1.1.1
mecab-python3==0.996.2
mypy==0.720
mypy-extensions==0.4.1
openpyxl==2.6.3
pygame==1.9.6
Pygments==2.4.2
pytz==2019.2
requests==2.22.0
soupsieve==1.9.3
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4
uritemplate==3.0.0
urllib3==1.25.3
